refactor: inputs, uncontrolled form
